// projects/ngx-forge-map/src/lib/model/interfaces/project.model.ts
export interface ProjectJSON {
  id: number;
  description: string;
  name: string;
  name_with_namespace: string;
  path: string;
  path_with_namespace: string;
  created_at: string;
  default_branch: string;
  tag_list: string[];
  topics: string[];
  ssh_url_to_repo: string;
  http_url_to_repo: string;
  web_url: string;
  readme_url: string;
  forks_count: number;
  avatar_url: undefined;
  star_count: number;
  last_activity_at: string;
  namespace: ProjectNamespaceJSON;
}

export interface ProjectNamespaceJSON {
  id: number;
  name: string;
  path: string;
  kind: string;
  full_path: string;
  parent_id: undefined;
  avatar_url: undefined;
  web_url: string;
}